fix(archive): 되쓰기 판정을 react-query 없는 순수 함수로 — 검증 게이트 복구
`mutations.spec.ts` 가 `@tanstack/react-query` 를 값으로 import 해서, #780 머지
(140ff618, 09-03 06:04 UTC) 이후 develop 의 검증 게이트가 계속 빨갰다. admin-web 은
npm workspace 가 아니고 루트 `npm ci` 는 `apps/admin-web/node_modules` 를 깔지 않으므로
CI 러너엔 그 패키지가 아예 없다 — 로컬엔 깔려 있어 작성 시점엔 초록으로 보였다.

판정이 걸린 병합 로직을 `write-back.ts` 의 `mergeSavedPage` 로 빼고 스펙을 그쪽으로
옮긴다. `writeBackSavedPage` 에는 캐시 키 배선만 남으며 시그니처·호출부는 그대로다.

형제 스펙(`products/queries.spec.ts`)은 같은 문제를 `jest.mock(…, { virtual: true })`
로 피하지만, 여기선 QueryClient 를 흉내 내는 대신 실제 병합 동작을 그대로 검증하는 쪽을
골랐다. 대신 «어느 캐시 키에 쓰는지» 는 스펙 밖의 한 줄로 남는다.

검증: `@tanstack/*` 해석을 막은 채(= CI 조건) 새 스펙 4건 통과 — 같은 조건에서 옛 스펙은
그대로 실패하는 것으로 재현 확인. 루트 `type-check` 에러 0, 루트 `jest` 에서 이 스위트 통과.
